% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_info}
\alias{card_info}
\title{Afficher une fiche}
\usage{
card_info(name, lang = "fr")
}
\arguments{
\item{name}{Identifiant de la fiche, par exemple `"VCN10"`.}

\item{lang}{`"fr"` ou `"en"`.}
}
\value{
La figure, invisible, telle qu'elle est affichée.
}
\description{
Afficher une fiche
}
